I hate Tcl.
Here's a macro that tries to figure out which version of Tcl braindamage
a user has inflicted on his system: vanilla, Sun, BSD; 7.5, 7.6, 8.0, 8.1.
Should be tweakable as long as people keep picking at it, I hope.

Relies on Tcl's tclConfig.sh script to get its information, which either makes
it massively more or less reliable than ad-hoc approaches.  It's not clear
which.