removed an unnecessary free before it can do any damage.
Note that because of my previous change a moment ago, callers to
appendstrlistpat (all in imapd.c, as far as I can tell) must allocate memory
before calling.  This looked to be cleaner and it was certainly easier to
change it this way then to keep the xstrdup in appendstrlistpat.