skiplist locking refactor
From Bron Gondwana <brong@fastmail.fm>:

John Capo discovered a bug whereby foreach wasn't updating the transaction
pointer before calling the handler function, so any updates during the foreach
would fail if (and only if) it was the first database function to use that
transaction.

He posted a trivial patch to the mailing list which fixed the bug but left
unreachable code.

I decided that what was actually needed was a refactor of the 4 subtly
different sets of locking on the access functions: fetch, foreach, store and
delete into a single locking function.

The attached match creates a new function lock_or_update(), and also always
uses malloc and makes newtxn a "creation is initialisation" function for the
transaction.  It's a much neater way to handle transactions, and significantly
clearer to read.