calculate a global "sessionid" for every connection or login
This patch exists to support our 'auditlog' infrastructure, which
allows tracing every action back to the login that caused it.  The
sessionid is returned in the login response or lmtp delivery response,
and is logged all the way back through the proxy/filter chain, so we
can follow every message end-to-end.

===================================================================