Check subscription database rather than explicit ADDSUB/DELSUB
We were seeing occasional subscription mismatches at FastMail,
and the root cause was traced back to the re-ordering of log
files.  All the ADDSUB commands would run first, then all the
DELSUB commands.

If a log file contained a DELSUB followed by an ADDSUB, they
would be re-ordered, and the subscription would exists on the
master but not the replica.

This patch changes the specialness of subscriptions, so that
instead of SUB and UNSUB events, it's just SUB events, and the
sync_client will check the subscription database to find what
the value should be, then tell the backend to make it so.