Use REG_STARTEND instead
The PCRE library supports using REG_STARTEND to specify the start and
end offsets, which the BSD library also supports, and it looks like
recent Linux libc as well.  So if we have REG_STARTEND, use that.

It appears the RX library doesn't support REG_STARTEND, so use its API
if we don't have REG_STARTEND but do have RX.

Finally, fall back to duplicating the entire string and using the
POSIX API directly.