Add MAX_MAILBOX_BUFFER with enough space for name manipulation
It appears to be possible to create a mailbox with a name so long
that you can't delete it.  That sucks!

This patch works around it by having a bigger buffer for all the
name manipulation that goes on behind the scenes.

Also, quota.root only needs to be a MAX_MAILBOX_NAME size in
reality, so make that use MAX_MAILBOX_BUFFER too.

And get rid of a bunch of random use of MAX_MAILBOX_PATH when
it's bogus.  Not to mention some fixed buffer lengths.