Write entire header during upgrade
This patch keeps a copy of the index header record written at the start
of the index upgrade so it can write it entirely again at the end rather
than seeking inside and writing just a part of the record.

This is to set up for CRC32 patches later, which will need to have the
complete header record on hand to recalculate.  Otherwise, it's just
neater anyway.