Don't abort BDB transaction on force delete of non-existant key
The check was being performed after the abort, causing mydelete
to return CYRUSDB_OK, but the transaction still to be aborted.
Oops.

Moved the check to directly after the db->del, which coincidentally
also removes the same bug from the !mytid path, which didn't even
check for force at all!