Replace ctype macros
... which all expect int argument in the range 0..255
or EOF (typically -1). I have left a few instances of isspace() and
friends where the input comes from fgetc() or prot_getc(), and
consequently is in the expected range.

Mostly automatic using find and sed, but I did go back and remove
casts to (unsigned char) [correct but redundant], and (int) [typically
incorrect if the source is signed char, probably put there to stop
the compiler moaning].

(thanks to David Carter)