Use correctly sized variable for recovery time
Global Recovery was being htonl'd into a time_t,
which is a 64 bit value on some platforms.

Depending on the host bit order, this might be
promoted into the last 4 bytes of an 8 bit value,
meaning that the bytes written to disk were actually
all zero.

Change it to a bit32 and give it a more descriptive
name.

Thanks to Michael Bacon <baconm@email.unc.edu> for
finding the bug.