Fork idled before cyrus_init()
We were getting:

* DBERROR db4: environment reference count went negative

Reliably in one random process when shutting down our cyrus
instances, and this is even with no BDB databases defined in
our config.  The cause was tracked to idled calling cyrus_done
in both the parent and child processes, but cyrus_init only in
the parent, hence closing the environment twice.

It's possible this also caused other issues for BDB, since it's
invalid usage of the API.

This change moves the fork to before cyrus_init(), meaning that
the environment is only ever initialised in the child process.