Track idling state and tell idled if we lose connection
We get a bunch of "process killed with signal 10" errors
after restarting one of our imap proxies, and it's because
idled goes a little crazy throwing SIGUSR1 and SIGUSR2 at
processids which have long since gone away.  It's like a
loaded gun that randomly shoots whatever process was unlucky
enough to get that PID.

Now, the whole design sucks really - but at least the worst
of it will be fixed by telling idled that we've idle_done()
when we shut_down(), no matter how we got there.