Add "-x" option to cyr_expire to disable expunge
Delayed expunge is great and all, but it causes a LOT of IO
because you need to stat every mailbox's meta files, and you
need to read all of every cyrus.expunge file.

That's not something you might want to do every time you want
to clean out the duplicate delivery database.

Change cyr_expire to add a "-x" option.  Ideally I'd like to
just have not specifying a "-X" mean don't do it, but I think
it might be too late for that.  It would make disks fill up
with junk at too many sites.