Fix obscure timezone parsing bug in getdatetime
One of the legacy formats (i.e. formats not in the current IMAP
RFC) accepted by getdatetime() allows clients to specify the
timezone as a three-letter abbreviation for one of several
North American timezones.  The offset calculated for the Standard
versions of these is correct, but the Daylight version is wrong
by two hours.  For example, Central Daylight Time is UTC-0500 but
getdatetime() thinks it's UTC-0700.