XFER out to version 2.2 and version 2.3
This defaults to exporting a version 6 cyrus.index file.  Version 6
was in all 2.2 since at least 2.2.10.

Version 7 was present in 2.3.0

Version 8 was present in 2.3.4

Version 9 was present in 2.3.7

Version 10 was present in 2.3.10 +

Version 11 was FastMail internal, it shouldn't exist in the wild any
more.

Version 12 was added in 2.4.0.

This code will detect version 10 or version 12 changeover points from
the banner of the destination IMAPd.  If you've supressed version
information in your banner then the mailbox will always transfer as
version 9, which will cause an upgrade at the far end and cost you
extra IO, but shouldn't break anything.