unit: add basic infrastructure for CUnit tests
A new subdirectory cunit/ contains the main routine of a driver program
for CUnit, and a Perl script to help at build time.  The driver is
built and run by the check: target.

The Makefile.in in each directory that has test source will need to add
a few lines to re-implement it's check: target, using the Perl script.

The build time perl script works on a project file called default.cunit
which is built from scratch at build time and contains a list of test
sources and test libraries.  The script has the following main functions:

--add-sources foo.c ...
	Add a set of test sources (sources and libraries can be
	safely added twice or more)
--add-libraries -L/usr/foo -lfoo libbar.a ...
	Add a set of test libraries
--generate-partial-makefile foo.mk
	generate a makefile fragment which contains rules to build
	object files from test source in this directory, without a
	final link
--generate-final-makefile foo.mk
	generate a makefile fragment which contains definitions of
	$CUNIT_OBJECTS and $CUNIT_DEPLIBS, which can be used to
	do a final link
--generate-makefile foo.mk
	generate a makefile fragment with both partial and final
	features, useful in the simple case where all the test source
	is in the same directory