unit: add commandline options to the unit program
New options are:
 -h    print help message
 -l    list all tests and exit
 -v    be more verbose

Any non-option arguments are treated as suite names or test names
(in the form SUITE:TEST) and only those suites or tests are executed.

The -v option makes CUnit print suite names and test names as it
prints them.  Two -v options shows any text emitted to syslog,
and also describes each successful CU_ASSERT*() as it happens.

The old behaviour was the equivalent of -v, so use that in the
check: target.