One copy of common signal processing is enough
Both libimap.a and libcyrus.a have had a copy of signal processing routines.
The one in imap/signals.c was unused, uncompiled yet confusing due to the
makefile snafu which was generating the right objects in the wrong directory.
It is now gone.

The differences between the two versions were:
  - volatile type for signal flag
  - SIGQUIT was registered explictly to avoid marking its handler
    as restartable.

Volatility of gotsignal has been restored.