Remove useless buffer arguments from charset.c
The functions charset_convert(), charset_decode_mimeheader() and
charset_decode_mimebody() all took char *buffer and length arguments.
These had the bizarre semantic that the buffer had to be malloc()ed so
that the code would realloc() it when it felt the need (rather than the
expected semantic for such an API where the conversion would fail if the
buffer limit was exceeded).  This worked only because every single
callsite passed NULL.

Remove the useless arguments.  Now charset_convert() explicitly
always returns a new string.