First cut at running a jailed Cyrus instance
Instead of operating on a separate Cyrus instance in the VM with known
connection parameters, provide the infrastructure to run a jailed Cyrus
instance.  Creates a skeleton directory, creates cyrus.conf and
imapd.conf files, sets up the mboxlist db with a single mbox, runs
reconstruct to create the metadata for that mbox, and finally starts
master.  Added new Simple test to exercise that.