unit: fix 2nd per-test timeout
The timeout code was working the first time and not the second or
subsequent times.  The reason was that the signal handler for SIGUSR1
was calling CU_FAIL_FATAL() which does a longjmp() back out to the main
loop.  As a result, the kernel never sees a sigreturn() syscall, which
normally happens when the signal handler returns out into the userspace
trampoline routine which called it, and so never knows that the signal
handler has finished.  This doesn't matter much, except that the
semantics of signal handlers registered with signal() are that the
signal being delivered is masked during the signal handler.  So when the
second SIGUSR1 was sent, it was masked and not delivered.

This could be fixed by calling sigreturn() in the signal handler, except
that this is not portable to Solaris.  A more portable fix is to use the
sigaction() call with the SA_NODEFER flag to turn off the masking of the
signal during delivery.  These signals are sent singly and 20 seconds
apart, so raciness is not an issue.

Also, be a little more pedantic about invalid values of some variables.