annotate: RFC5257 ANNOTATION Interaction with COPY
At least for the case of COPYing from a local mailbox to another local mailbox.

To support this, generalise annotatemore_rename() to _annotate_rewrite() which

- can do either renames or copies
- can change uid

Reimplement annotatemore_rename() and annotatemore_delete() in terms of
it, and add annotate_msg_copy() which uses it to handles the annotation
side effects of COPYing a message from one folder to another.