Generalised genmail3.pl
Added SequenceGenerator.pm which derives from Generator and just
generates a sequence of messages at fixed time offsets.  This is
the data pattern originally used for the pop3showafter tests, and the
loop was originally in genmail3.pl.

Generalised genmail3.pl to do sequence (default) or threaded (option
--threaded) message generation.  Also, genmail3.pl generates to an IMAP
folder by default, unless given a URI with -u or the name of an mbox
file to create.

Added a bunch of commandline options to genmail3.pl to allow changing
the IMAP parameters, such as --host --port --user --password --folder.