Bug #3453: use direct access to subs DB
This is instead of the super-buggy dance from internal namespace
to external and then back again which you get using mboxlist_findsub.

Adds a new API "mboxlist_allsubs" which takes a userid.  I'm seriously
tempted to add the same API for the mailboxes as used by sync_client -u,
but not just yet.