Run all utilities via _fork_utility()
An ancestor of this commit originally changed the code which ran
ctl_mboxlist and reconstruct to go through _fork_utility() so that the
environement and other tweaks applied to child processes in that
function are always applied.  That code is gone since commit "instance:
create a special "adminimap" instance that's used for creating users",
but some code cleanups of _fork_utility() are still useful.

In particular, die with a useful message when a process started by
_fork_utility() exits with a non-zero exit code, or is terminated by a
signal.  This ensures the test fails and there's enough information in
the failure message to track down why.