annotate: revise explicit transaction control API
Change the transaction control API recently added by Ken Murchison.

The annotate module manages a single file-scope struct txn.  This txn is
not exported outside the annotate.c, and the references to it in the API
are removed (they were always being passed NULL anyway).

The functions annotatemore_commit() and annotatemore_abort() now take no
arguments and work on the internal txn.  A function annotatemore_begin()
is added to explicitly begin a new txn (as there is now no way to do
this implicitly by passing a non-NULL txn**).

The behaviour of annotatemore_store() is changed to require that a
transaction be open.  All callers of it now call _begin()/_commit().

These changes simplifies some internal logic, and also allow multiple
annotatemore_store() updates to be made in a single transaction (useful
for STORE ANNOTATION with multiple messages).