sync_client - only sync 1000 mailboxes at once
This avoids the worst of a horrid O(n^2) algorithm with
name list scanning.  It's still pretty bad reading the
log file in the first place, but at least it doesn't
loop the whole list for every mailbox multiple times
this way.

Changed a log with 300,000 mailboxes in it to run in
under 1/2 hour, when it had previously been 3 hours in
before I killed it and added this patch!