annotate: fix FMR after annotatemore_msg_lookup().
This function fills in the given struct buf with a read-only copy of the
data buffer which belongs to the cyrusdb layer, before calling
annotate_putdb.  If the annotation is per-message, and we're being
called outside an explicit transaction, that putdb call will drop the
final reference and free the db including the data buffer we've just
saved a pointer to.

Fix this by using buf_cstring() to force the struct buf to have a copy
of the data rather than a pointer into the db layer's buffer.