Fix indentation in imapd.conf.5
There are several multi-paragraph comments in lib/imapoptions delineated
by the .PP macro, and one definition list implemented with .IP.  When
converted to man page format, these are badly indented.  Work around
this in the Perl script that generates the manpage by surrounding
multi-line descriptions of options with a .RS/.RE pair.