annotate: better checking of defined attrib-names
Improve the checking of the specified values for the "attrib-names"
field in the annotations definitions file.  We were already ignoring
deprecated values with a warning.  Now also silently ignore "size",
"size.priv" and "size.shared" and instead calculate these from the
presence or absence of "value.whichever", so that the server obeys
RFC 5257 regardless of what was typed into the annotation definitions
file.