Revert merge of stable branch fix for Bug#3489
This use-after-free bug was fixed on the 2.4. stable branch a couple of
days ago.  The fix was then merged back to master.

However the bug had already been fixed in the master branch back in
January as a completely unintended side effect of the data structure
cleanup work done in commit 310f1c7 "Replace qr_path with strarray_t".

The code resulting from the merge initially didn't compile, and then
when made to compile introduced a double-free regression.  This commit
reverts that merge, undoing the effects of commits:

79499c4 quotalegacy API usage fixup for master
4aa2b60 Bug #3489 - fix quotalegacy handing of prefixed foreach