Fix message cache item offset calculation.
When the cache contents are calculated, cache item offsets are also
calculated.  That calculation was broken by me in commit "Fix extra NULs
in cache items", such that the offset pointed to the 4-byte character
count at the start of each item instead of to the first character of
the item data.

Note that normally the offsets calculated here are ignored; the cache
data is saved in its entirety without reference to them, and the correct
values are calculated when the cache data is loaded again.  However in
the "future" branch there is some conversations code which uses the
cache item offsets before they go out to the file, and that code was
quite subtly broken.

Found by running Cassandane test with Valgrind.