Fix possible UMRs due to parse_cached_envelope()
This function fills in an array of char*s but provides no way for the
caller to know whether it has filled them all in or not.  Worse, none of
the callers initialise the arrays they pass in and several of the
callers use elements of those arrays as if they were initialised.

The smallest impact fix here is to initialise the array to NULLs as the
first step of parse_cached_envelope().

Such UMRs were actually seen as a cascading failure from the botched
cache item offset calculation bug, while running Cassandane tests with
Valgrind.