annotate: don't fatal() when reading definitions
Calling fatal() as an error-handling mechanism when parsing the
annotation definitions file has proved to be wildly dangerous (by
causing a startup loop in the master program).

So, rewrite the parsing code to handle errors in the file by emitting
useful messages to syslog and struggling on.  Messages include the
filename, line number, column number, and usually the token near which
the error occurred.

Also rewrote the awful open-coded string scanning code to use the new
tok_t structure for tokenising.