add username named arg to start_replicated_pair()
allowing the caller to specify the username by which the stores login to
both the master and the replica instances.  This allows creating a
replicated pair connected as the admin user, which will be handy for
Quota tests.

As a side effect, the Config object which was previously an optional
position parameter is now a named parameter.