Add tests for the Cyrus annotator daemon
The program utils/annotator.pl is an annotator daemon which uses the
Cyrus::Annotator::Daemon class.  It parses the message body as a
sequence of test commands and executes them; this allows the Cassandane
test case to drive it's behaviour remotely.

Cassandane/Cyrus/Annotator.pm tests setting annotations and flags on a
locally delivered message.