Refactor Quota tests using utility functions
which reduces the verbosity of doing and checking quota operations, thus
allowing the recently added _late variants of several tests to be split
out as their own tests without any 'if ($late)' stuff.  I really really
prefer to have each top level test function to be as simple and linear
as possible, in particular with a single execution path.