"flat" DB backend is binary-safe for keys & data
at least for the specific case of embedded \0 \t and \n which were
definitely broken.  Using the character 0xff as an escape character,
because it's an invalid UTF-8 octet and a very unusual latin1 character
and so unlikely to be present in existing textual data.  Existing binary
data was of course already broken.

This should now allow the "flat" format to be used as a safe exchange
medium for all Cyrus databases, even the annotations db which uses
embedded \0 in the key and the data.