Add a regression test for BZ2686
which calls cvt_cyrusdb to convert an annotations db from skiplist to
flat file format and back again, then checks that the new database
contains all the data the original one did.  This used to fail because
the flat file format could not handled certain binary characters
embedded in keys or data.