Fix LSUB as found by Cassandane and bisect
This bug was introduced in 7c1ce5ba413f8f87ee163930f6afa91dfd5c81ab.

The change in behaviour is that cyrusdb_flat now returns data == NULL
when datalen == 0.  Before it returned a pointer to a \0.
mboxlist_findsub and friends were checking that data != NULL when
they didn't need to, all they needed to check was a zero return from
fetch, because it returns CYRUSDB_NOTFOUND if there's no match.