[IRIS-883] Don't spuriously require the 'a' right
on certain builtin entries.  This represents a change in behaviour from
Cyrus 2.4.  Although it's due to fixing what could be called a bug in
the old code (whereby some of the static annotate_st_entry structures
had .acl=ACL_ADMIN despite that field never being checked for entries
which did not go to the database), we aim to be compatible with 2.4 in
these grey areas.