sync: don't parse message files while holding locks
Trying to move a large user I noticed that the global
annotations.db lock was held for ages while we did
guid checks of every single message file.  Oops!

Fixed to not do that any more.  Instead, we check
while uploading it.  We'll bail regardless, but at
least it won't be under a lock.