lib/mkchartable.pl move output from stderr to stdout
lib/mkchartable.pl is used to generate lib/chartable.c on stdout .  So far
it pritend status messages to stderr.  Status messages do not belong to
stderr, but to stdout.

The applied patch adds one parameter to mkchartable.pl:
  -- output / -o , specifying to which file the output shall be stored,
and  fixes Makefile.in to make use of the -o parameter.