Add another regression test for duplicate mode
It turns out there was a bug in commit

"[IRIS-723] use mailbox uniqueid for duplicates"

which added the imapd.conf option "duplicate_mailbox_mode".  With
the (non-default) value "uniqueid", attempting to deliver a message via
LMTP to a non-existant mailbox would crash lmtpd, instead of sensibly
defaulting to INBOX.

The new test Delivery.duplicate_suppression_on_uniqueid_badmbox
tickles that corner case, and is a regression test for the fix in
commit "lmtp: don't crash on non-existent mailbox append attempt".