perl/Makefile*: enable building perl modules out of srcdir
Adds ability to build the perl modules in a directory different from
the source directory.  This statement is actually wrong, since the
automatic variables $* in Makefiles in Old-Fashioned Suffix Rules
(as described the the make manual) do not expand to contain the path
of the (input)-file, just its name.  Hence, if ExtUtils::MakeMaker
used $< instead of $* in the generated Makefile, it would be possible
to build the complete cyrus-imapd in a directory, different from the
source directory.