Avoid cascading failures during instance setup
If a test failed during the creation of an instance, an exception
describing the problem might be thrown before the instance's conf/cores/
or vglogs/ directories exist.  When this happens the instance cleanup
code gets upset about the inability to check for Valgrind log files and
the like, and throws another exception which hides the original and
useful exception.  Avoid this by returning early from
_check_valgrind_logs() and _check_cores() if the respective directories
are missing entirely.