Push more initialisation into Instance::start
and out of the Instance constructor, because start() is run in the
Worker process in parallel mode but the constructor is run in parent
process.  This was interfering with the generation of instance basedirs
for those few tests (like Bug3072) which still create Instances in their
constructors.

This means that $instance->{basedir} isn't defined until after start()
or stop(), which turns out to be a little bit tricky.

Also, delay the pwcheck setup until start().  This means that there are
no Instance methods called in the constructor anymore and the very last
thing it does is 'return bless'.

Also, fix a race condition in the creation of basedirs, which shouldn't
matter unless two Cassandanes are run at once, but at least it's done
properly now.