master: fix SIGCHLD vs select() race
Cyrus' master process is designed to rely entirely on delivery of
SIGCHLD to detect when child processes have died.  It has no other way.
In particular, master cannot detect an EOF on the status pipe used to
communicate with children, as this status pipe is shared between all
children and the number of writer fd on the pipe never falls to zero.

Detection of SIGCHLD occurs in the traditional way, with a signal
handler which sets a global volatile flag, and a main loop where
select() returns early with EINTR when a signal is delivered, forcing a
check of the global flag and a call to waitpid().

This arrangement suffers from the traditional UNIX race condition.  If
SIGCHLD is delivered *before* the select() call starts, the application
doesn't notice until select() returns for some other reason.  This race
has been known to deadlock servers, but it's impact in Cyrus is more
limited because master wakes itself up every 10 seconds anyway.

The upshot is that the time taken between a child process dying and
master reaping, may racily jump to 10 seconds instead of the usual
number which is in the low millisecond range.  The race window to
trigger the bug is pretty small, but it turns out it can be made to
happen reliably by either:

a) inserting a 20 millisecond nanosleep() before the select(), or

b) running master under Valgrind (but not on my laptop).

The solution to this well-known race condition is also well-known and
was standardised by POSIX: pselect().  This variant of the select()
system call takes an extra argument which is the signal mask to apply
during the pselect() call.  Master now blocks SIGCHLD at initialisation
time, and calls pselect() with a sigmask which is calculated to be the
same as the existing sigmask but without SIGCHLD.  Thus, SIGCHLD is only
ever delivered while master is in pselect().

Of course we do some autoconf magic to ensure we only try to use
pselect() on platforms where it exists.