master: enforce maxforkrate at high granularity
Enforcement of maxforkrate was broken because of a thundering herd
effect that occurred when the logic decided that the limit should be
removed.  At that point, *all* the pending connections which had been
ignored are immediately serviced in a busy loop as we go around the
pselect() loop.  This is because the actual fork rate is not
re-estimated until another "forkrate interval" of almost 2 seconds has
passed, so the forks due to these pending connections are not counted
until it's too late.

The solution is to re-estimate the fork rate every time
service_is_fork_limited() is called instead of only every 2 seconds.
Also, we reduce the ignore time from 2 seconds to 0.4 seconds, using our
new ability to schedule events at subsecond granularity.