autobuild: detect manual vs Jenkins builds better
Instead of relying on a --manual argument, use the existing check for
the existance of the various Jenkins environment variables to decide
whether we're supposed to be running under Jenkins or not.

Also, detect whether to do code coverage by testing for the presence of
ggcov tools in $PATH rather than hardcoding paths to ggcov executables
which worked for me only.

Finally, print a more useful summary of build parameters and environment
at the start.

autobuild.sh should be usable manually by developers now.