Bug 3648 - master handles SIGQUIT synchronously
The SIGQUIT handler now just sets a global volatile flag, which is
checked in the main loop.  The work of sending on SIGQUIT to the
children is now done in a new function begin_shutdown().  Instead of
ignoring SIGQUIT, we keep handling it but don't do anything if
begin_shutdown() is called twice.

Also, to close a race where the re-sent SIGQUIT is ignored by a child
which is between fork() and exec() spawned just before the master
receives SIGQUIT, we explicitly set the handler for SIGQUIT back to
SIG_DFL before unblocking the signal.  This should stop the leakage of a
lemming process from test Master.prefork.