master: don't busyloop after SIGQUIT
After shutdown is started, no new events are scheduled.  Usually this
mean that a single periodic janitor event remains on the queue, with a
scheduled time that soon moves from the near future into the past.  The
timeout value calculated for select() is based on this, and soon goes to
{0,0} which results in master busy looping through select() until the
children eventually die (or for ever, before the previous bug fix).