Bug 3100 master: handle SIGTERM synchronously
so now all the signal handlers just set global volatile flags, and none
of them do anything fancy.  This should help remove all sorts of awful
race conditions.

Removed all the semantic differences between SIGQUIT on the one hand and
SIGTERM/SIGINT on the other.  Previously, SIGQUIT meant "send a signal
to all child processes and wait for them to shut down gracefully, then
shut down gracefully yourself", whereas SIGTERM/SIGINT meant "send a
signal and then just exit, and never mind the child processes".

Now we should have enough bug fixes in place that master can reliably
wait for it's children without busy looping or leaking child processes,
so master always waits for its children to shut down gracefully
regardless of which signal starts shutdown.

Also, inexplicably the SNMP code was gracefully shut down only on
SIGTERM/SIGINT, now that happens for SIGQUIT too.

Also, the signal sent to child processes is now SIGTERM not SIGQUIT.
The main difference is that the kernel default action for SIGQUIT is to
terminate and dump core, whereas SIGTERM just terminates.  There is a
small window during child startup (between the execv() and the child
calling sigaction() to install it's own signal handlers) when the child
is necessarily using the kernel default actions.  If master sends
SIGQUIT to a child in this window, it dumps an unsightly and unhelpful
core.  The Master.prefork tests were hitting this race window (as well
as the one between the fork() and the exec()).

With this change, the Master tests in Cassandane are now stable in my
environment :)