idle: centralise AF_UNIX socket creation
and address construction.  One significant change is that we make the
client create a bound socket with a per-client pathname for its
communication.  This will make it easier (in a later commit) for the
server to reply to that socket instead of having to use clunky and
dangerous async signals.