idle: don't use signals, use AF_UNIX dgrams
Communications back from idled to imapds are via a message sent on the
AF_UNIX socket.  The IDLE command is implemented as a select() loop, and
there's absolutely nothing that needs to be done in signal handler
context.  Best of all, no more unexpected delivery of SIGUSR1 or
SIGUSER2, assassinating innocent bystander processes.