imap/Makefile.in: remove linking with tls.o, when libimap.a linked
tls.o is included in LOBJS,
LOBJS is included in libimap.a,
so whenever libimap.a is used, tls.o is available.

However some targets link both tls.o and libimap.a, this commit removes
in such cases tls.o, and adds the rest of the dependancies and the command
in alphabetical order.