configure.ac: remove check for libfl
libfl is installed by flex and provides the main and yywrap functions,
if the generated lexers don't have their own implementations.  However,
in sieve/libsieve.a there is no place for main() function and both
sieve/addr-lex.l and sieve/sieve-lex.l declare %option noyywrap, so
this library is not necessary.  The point is, that the check adds the
library automatically to LIBS, so at end there is main() in IMAP_LIBS,
which is messy.