lib/Makefile.am: move strarray from libcyrus.a to libcyrus_min.a
master/Makefile.in: remove dependance on lib/lock_@WITH_LOCK@.c

lib/strarray.c is not included in libcyrus_min, but in
lib/libcyrus.a . As master/master uses lib/strarray.c and libcyrus_min,
and all other programs using libcyrus_min.a use also libcyrus.a,
strarray is moved from libcyrus.a to libcyrus_min.a and the explicit
linking with strarray.a from master/Makefile.in is removed.

The file lib/lock_@WITH_LOCK@.c is anyway included in lib/libcyrus_min.a,
master links with libcyrus_min, so it is not necessary to request linking
with lock_@WITH_LOCK@.c once again.