Lsub tests create a user with multi-level folders
using the syntax

$inst->create_user('user', subdirs => [ 'sub.dir' ] )

Recent changes to support handling unixhierarchysep cleanly, which threw
an exception if the incoming folder name contained the hierarchy
separator, broke this usage without providing a replacement.

Add support for specifying new folder names to Instance::create_user()
as arrays of strings, rather than a single string containing the default
hierarchy separator, and change the Lsub tests to use this.  The new
syntax looks like

$inst->create_user('user', subdirs => [ ['sub','dir'] ];

This should fix the broken Jenkins builds #507 (and probably #508 if I
don't check this in fast enough).