rehash: consistent messages and better options
Can now pass '-n' to not make any changes, '-v' for verbose output,
and -f|F and -i|I to change the hashing options.  If no hashing
options are given on the command line, they are read from imapd.conf.

There are some risks with running this command with 1 char named
top level mailboxes or users - so don't do that.