Bug 3652 - set Perl search paths correctly
Finding Cyrus Perl modules in Perl code is not trivial.

 - There are up to 4 possible installation directories: (site,vendor) X
   (arch,lib).

 - Installation directories are derived in unobvious ways from the
   configuration of the build-time Perl, not the run-time Perl, and are
   different on different OSes.

 - Cassandane runs Cyrus executables from an installation staging
   directory, not a full install, whose value is not predictable at
   build time.

 - The code to work this out is non-trivial and should be centralised,
   yet it cannot be put in a module because it's needed to...find
   modules.

So this latest approach does most of the work at build time, in a Perl
script which is used as a pre-install filter for Perl scripts.  It
calculates the part of the include paths which can be calculated at
build-time and emits Perl code which does the remainder at run-time.
This boilerplate inserted by the filter before the first "use" statement
in the script.

Also, changed the way cyradm.sh bootstraps from shell to Perl to be
consistent with installsieve and sieveshell.

The rest of the commit is mostly build-time glue to get the arguments
for the pre-install script into the actual Makefiles that need them.
MakeMaker makes this unobvious, of course.

Thanks to Rob Mueller for discussions and review.