Makefile.am: don't use per target (LDAP_)CPPFLAGS
When per target CPPFLAGS are used, e.g. preprocessing ptloader with
-I${LDAP_CPPFLAGS}, Automake assumes, that the same source files,
ptclient_ptloader_SOURCES, can be compiled once with the general AM_CPPFLAGS
and once with the per target ptclient_ptloader_CPPFLAGS.  To avoid
confusions, Automake creates then an .o file with very-long name, stating
explicitly in the name of the .o file, for which target the .o file will
be used.  These long names of .o files make the Makefile.in file longer, by
adding extra explicit rules for the ptloaded-target.

Having global single AM_CPPFLAGS for all targets simplifies the
Makefile.in and leads to .o files with shorter names.