Bug 3655 - Make quota -f more efficient
Replace the OPT_MAILBOX_QUOTA_SCANNED mailbox flag with the new quota
scanset API, and move that check down into quota_update_useds() which
now takes a mailbox name instead of an is_scanned flag.

This allows quota -f to proceed with the same race protection as before
but with far fewer fsync()s to clear and then set the flag on every
mailbox.  Instead we read and write an additional record in the quota
db, which already has a open RW transaction at all the points where we
use the new API.