Remove 256 byte limit on cyrus.conf arg values
by pre-splitting the line into arguments when the line is parsed from
the file and not later on demand.  The limits are now

- lines must be at most 4095 bytes long
- lines must contain at most 64 arguments

This also removes an aliasing problem, where the following

    foo cmd="/usr/bin/foo --listen=bar --fnarp" listen=quux

would do the wrong thing when asked for the value of the "listen" arg.

The 256 byte limit was very annoying when running Cassandane with
--valgrind, which would generate "cmd" arguments longer than that.