Makefile.am: move the defintion of COM_ERR_LIB here
When the implementation of libcom_err.a, distributed with cyrus-imapd is
used, com_err/et/libcom_err.a shall be build, before the programs depending
on it are linked with it (as linking does not work otherwise).  By moving
the definition of COM_ERR_LIB from configure.ac to Makefile.am (for this
case) Automake knows, that the value of COM_ERR_LIB might be
com_err/et/libcom_err.a and Automake can and does create the necessary
dependencies.  In turn, when building something that depends on the bundled
libcom_err.a, it is ensured, that com_err/et/libcom_err.a is built on the
right time.