quota: use dlist for storage and replace 'quota -f' again
This time we store a 'SCAN' item in the quota file, like
so - a regular file might be:

%(S (123456 1000000) M (500) AS (0))

And during the scan, might be:

%(S (123456 1000000) M (500) AS (0) SCAN %(MBOX "user.foo.Sent Mail" S 85764 M 400 AS 0))

Everything which updates quotas can just use the name of MBOX
in the 'SCAN' key to determine if it should also update the
scanuseds[] array - if the mailbox is sort-order less than or
equal to, then it should.

This is race-free except for creation of a folder AFTER the
last folder name during a quota -f run which covers multiple
folders.  This could be fixed by putting mailbox create/delete
under the $QUOTACHANGE lock - but it means nobody can create
or delete any folder while 'quota -f' is running, instead of
just blocking the creation and deletion of quotaroots.